requirements.txt
asgiref==3.3.1
beautifulsoup4==4.9.3
certifi==2020.12.5
chardet==4.0.0
Django==3.1.7
django-environ==0.4.5
django-livereload-server==0.3.2
djangorestframework==3.12.2
idna==2.10
pytz==2021.1
requests==2.25.1
six==1.15.0
soupsieve==2.2
sqlparse==0.4.1
tornado==6.1
urllib3==1.26.3
